refactor: rename TenderItemPage to TenderPage
