--- src/components/TenderItemPage/index.js
@@ -1,1 +1,0 @@
-export * from './TenderItemPage';
