+++ src/components/TenderPage/TenderPage.styles.js
