+++ src/components/index.js
@@ -1,4 +1,4 @@
 export * from './Main';
 export * from './TenderItem';
-export * from './TenderItemPage';
+export * from './TenderPage';
 export * from './TendersList';
